backend/bot/requirements.txt
python-telegram-bot==21.9
httpx==0.27.0
python-dotenv==1.0.0
cachetools==5.3.3